raf: make waiting reader and writer increase and descrease atomically.

+++ include/lock/rwlock.h
@@ -19,8 +19,8 @@ typedef struct RWLockEntry {
     volatile s_lock content_lock;       /* Content spinlock. */
     volatile s_lock sync_lock;          /* Sync spinlock. */
     List *owner;                        /* Content lock owner rocesses. */
-    volatile uint64_t waiting_reader;   /* Waiting readers number. */
-    volatile uint64_t waiting_writer;   /* Waiting writers number. */
+    volatile int waiting_reader;        /* Waiting readers number. */
+    volatile int waiting_writer;        /* Waiting writers number. */
 } RWLockEntry;
 
 #define NOT_INIT_LOCK(entry) \
